+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+0.4.7.2
+-------------------------------------------------
 * Added cassava's `ToNamedRecord`, `ToRecord`, `FromNamedRecord` and `FromRecord` instances
 * Added `KeyIs` and `ValueIs`
 * Added `FromJSON` and `ToJSON` instances for `(:*)`